+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Bug fixes:
 * Fix the label of the first location reported by `Thread#backtrace_locations` (#2229).
 * Fix `Thread.handle_interrupt` to defer non-pure interrupts until the end of the `handle_interrupt` block (#2219).
 * Clear and restore errinfo on entry and normal return from methods in C extensions (#2227).
+* Fix extra whitespace in squiggly heredoc with escaped newline (#2238, @wildmaples and @norswap).
 
 Compatibility: